+++ OUDS/Core/Themes/Orange/Sources/Providers/ComponentTokens/OrangeThemeTextInputComponentTokensProvider.swift
@@ -16,46 +16,48 @@ import OUDSFoundations
 
 // swiftlint:disable type_name
 
-/// A class which wraps all **component  tokens of input text** for *input text* objects like `OUDSInputText`.
+/// A class which wraps all **component  tokens of input text** for *input text* objects like `OUDSTextInput`.
 /// Contains also references to semantic tokens providers so as to be able to use them to define the component tokens.
-/// This provider should be integrated as a `AllInputTextComponentTokensProvider` implementation inside `OUDSTheme` so as to provide
+/// This provider should be integrated as a `AllTextInputComponentTokensProvider` implementation inside `OUDSTheme` so as to provide
 /// all tokens to the users. It helps users to override some of the tokens and assign them to an `OUDSTheme` implementation to use.
-/// Custom themes can use subclass of ``OrangeThemeInputTextComponentTokensProvider`` and apply the provider they need.
-/// It implements also the protocol `InouttextComponentTokens` so as to expose the component tokens for *input text* through any `OUDSTheme`.
-/// *Inout text* components tokens are defined with raw and semantic tokens of sizes (from `AllSizeSemanticTokensProvider`), colors (`AllColorSemanticTokensProvider`)
+/// Custom themes can use subclass of ``OrangeThemeTextInputComponentTokensProvider`` and apply the provider they need.
+/// It implements also the protocol `TextInputComponentTokens` so as to expose the component tokens for *input text* through any `OUDSTheme`.
+/// *Inout text* components tokens are defined with raw and semantic tokens of sizes (from `AllSizeSemanticTokensProvider`), colors (`AllColorSemanticTokensProvider`), borcers (from `AllBorderSemanticTokensProvider`)
 /// and spaces (from `AllSpaceSemanticTokensProvider`).
 ///
 /// ```swift
 ///     // Define your own provider for input text component tokens
 ///     // by inheriting from existing provider
-///     class CustomInpuTextComponentTokensProvider: OrangeThemeInpuTextComponentTokensProvider {
+///     class CustomTextInputComponentTokensProvider: OrangeThemeTextInputComponentTokensProvider {
 ///
-///         // Then override the input text component tokens you want.
+///         // Then override the text input component tokens you want.
 ///
-///         override var inputTextSizeMaxWidth: SizeSemanticToken { DimensionRawTokens.dimension3000 }
+///         override var textInputSizeMaxWidth: SizeSemanticToken { DimensionRawTokens.dimension3000 }
 ///
-///         override var inputTextSpacePaddingBlock: SpaceSemanticToken { spaces.spacePaddingInlineSm }
+///         override var textInputSpacePaddingBlockDefault: SpaceSemanticToken { spaces.spacePaddingInlineSm }
 ///
-///         override  var inputTextColorBgDefaultFocus: MultipleColorSemanticTokens { colors.colorRepositoryOpacityBlackHigher }
+///         override var textInputColorBorderEnabled: MultipleColorSemanticTokens { colors.colorRepositoryOpacityBlackHigher }
+///
+///         override var textInputBorderWidthDefault: BorderWidthSemanticToken { borders.borderWidthMedium}
 ///
 ///         // ...
 ///     }
 ///
 ///     // Or define your own provider from scratch
-///     class CustomInpuTextComponentTokensProvider: InpuTextComponentTokens {
+///     class CustomTextInputComponentTokensProvider: TextInputComponentTokens {
 ///
 ///         // And implement hunders of tokens.
 ///         // You are allowed to give semantic tokens providers if you want to define values.
 ///     }
 /// ```
 ///
-/// Then, you can give this `CustomInpuTextComponentTokensProvider` to your own theme implementation:
+/// Then, you can give this `CustomTextInputComponentTokensProvider` to your own theme implementation:
 ///
 /// ```swift
 /// class LocalTheme: OrangeTheme {
 ///
 ///     override init() {
-///         super.init(inputText: CustomInputTextComponentTokensProvider())
+///         super.init(textInput: CustomTextInputComponentTokensProvider())
 ///     }
 /// }
 /// ```
@@ -63,7 +65,7 @@ import OUDSFoundations
 /// or to an already existing theme for example:
 ///
 /// ```swift
-///     OrangeTheme(inputText: CustomInputTextComponentTokensProvider())
+///     OrangeTheme(textInput: CustomTextInputComponentTokensProvider())
 /// ```
 ///
 /// It is also possible to use your own semantic tokens providers for this component tokens providers:
@@ -72,48 +74,55 @@ import OUDSFoundations
 ///     // Uses by default here:
 ///     // - OrangeThemeSizeSemanticTokensProvider for sizes
 ///     // - OrangeThemeColorSemanticTokensProvider for colors
+///     // - OrangeThemeBorderSemanticTokensProvider for borders
 ///     // - OrangeThemeSpaceSemanticTokensProvider for spaces
-///     let inputTextComponentTokensProvider = OrangeThemeInputTextComponentTokensProvider()
+///     let inputTextComponentTokensProvider = OrangeThemeTextInputComponentTokensProvider()
 ///
 ///     // Or use your own size, color and space semantic tokens providers (or only some)
-///     let inputTextComponentTokensProvider = OrangeThemeInputTextComponentTokensProvider(
+///     let inputTextComponentTokensProvider = OrangeThemeTextInputComponentTokensProvider(
 ///                                                 sizes: CustomSizeSemanticTokensProvider(),
 ///                                                 colors: CustomColorSemanticTokensProvider(),
 ///                                                 space: CustomSpaceSemanticTokensProvider())
 /// ```
 ///
-/// - Since: 0.10.0
-open class OrangeThemeInputTextComponentTokensProvider: AllInputTextComponentTokensProvider {
+/// - Since: 0.17.0
+open class OrangeThemeTextInputComponentTokensProvider: AllTextInputComponentTokensProvider {
 
-    /// Provider of size semantic tokens to use for button sizes
+    /// Provider of size semantic tokens to use for text input sizes
     public let sizes: AllSizeSemanticTokensProvider
 
-    /// Provider of color semantic tokens to use for button colors
+    /// Provider of borders semantic tokens to use for text input colors
+    public let borders: AllBorderSemanticTokensProvider
+
+    /// Provider of color semantic tokens to use for text input colors
     public let colors: AllColorSemanticTokensProvider
 
-    /// Provider of spaces semantic tokens to use for button spaces
+    /// Provider of spaces semantic tokens to use for text input spaces
     public let spaces: AllSpaceSemanticTokensProvider
 
     #if DEBUG
     private nonisolated(unsafe) static var instanceCount: Int = 0
     #endif
 
-    /// Defines a provider of component tokens dedicated to `OUDSButton`
+    /// Defines a provider of component tokens dedicated to `OUDSTextInput`
     /// - Parameters:
     ///    - sizes: Provider for size semantic tokens. If nil, a default one will be used (``OrangeThemeSizeSemanticTokensProvider``)
+    ///    - borders: Provider for border semantic tokens. If nil, a default one will be used (``OrangeThemeBorderSemanticTokensProvider``)
     ///    - colors: Provider for color semantic tokens. If nil, a default one will be used (``OrangeThemeColorSemanticTokensProvider``)
     ///    - spaces: Provider for space semantic tokens. If nil, a default one will be used (``OrangeThemeSpaceSemanticTokensProvider``)
     public init(sizes: AllSizeSemanticTokensProvider? = nil,
+                borders: AllBorderSemanticTokensProvider? = nil,
                 colors: AllColorSemanticTokensProvider? = nil,
                 spaces: AllSpaceSemanticTokensProvider? = nil)
     {
-        OL.debug("Init of OrangeThemeInputTextComponentTokensProvider")
+        OL.debug("Init of OrangeThemeTextInputComponentTokensProvider")
         self.sizes = (sizes ?? OrangeThemeSizeSemanticTokensProvider())
+        self.borders = (borders ?? OrangeThemeBorderSemanticTokensProvider())
         self.colors = (colors ?? OrangeThemeColorSemanticTokensProvider())
         self.spaces = (spaces ?? OrangeThemeSpaceSemanticTokensProvider())
         #if DEBUG
         Self.instanceCount++
-        checkInstances(count: Self.instanceCount, for: "OrangeThemeInputTextComponentTokensProvider")
+        checkInstances(count: Self.instanceCount, for: "OrangeThemeTextInputComponentTokensProvider")
         #endif
     }
 
@@ -125,8 +134,8 @@ open class OrangeThemeInputTextComponentTokensProvider: AllInputTextComponentTok
 
     // ଘ( ･ω･)_/ﾟ･:*:･｡☆
     // Note: So as to help the integration of generated code produced by the tokenator
-    // the implemention of InputTextComponentTokens is not here but in Core/Themes/Orange/Values/ComponentTokens/OrangeTheme+InputTextComponentTokens.swift
-    // This declaration of OrangeThemeInputTextComponentTokensProvider is here also to allow to write documentation.
+    // the implemention of TextInputComponentTokens is not here but in Core/Themes/Orange/Values/ComponentTokens/OrangeTheme+TextInputComponentTokens.swift
+    // This declaration of OrangeThemeTextInputComponentTokensProvider is here also to allow to write documentation.
 }
 
 // swiftlint:enable type_name