+++ OUDS/Core/Themes/OrangeBusinessTools/Tests/OrangeBusinessToolsThemeTests.swift
@@ -11,7 +11,7 @@
 // Software description: A SwiftUI components library with code examples for Orange Unified Design System
 //
 
-import OUDSThemesOrangeBusinessTools
+@testable import OUDSThemesOrangeBusinessTools
 import Testing
 
 // swiftlint:disable force_cast