+++ OUDS/Core/Themes/OrangeInverse/Sources/Providers/ComponentTokens/OrangeInverseThemeInputTextComponentTokensProvider.swift
@@ -16,24 +16,27 @@ import OUDSFoundations
 
 // swiftlint:disable type_name
 
-/// A class which wraps all **component  tokens of input text** for *input text* objects like `OUDSInputText`.
+/// A class which wraps all **component  tokens of input text** for *input text* objects like `OUDSTextInput`.
 /// Contains also references to semantic tokens providers so as to be able to use them to define the component tokens.
-/// This provider should be integrated as a `AllInputTextComponentTokensProvider` implementation inside `OUDSTheme` so as to provide
+/// This provider should be integrated as a `AllTextInputComponentTokensProvider` implementation inside `OUDSTheme` so as to provide
 /// all tokens to the users. It helps users to override some of the tokens and assign them to an `OUDSTheme` implementation to use.
 /// It implements also the protocol `InouttextComponentTokens` so as to expose the component tokens for *input text* through any `OUDSTheme`.
 /// *Inout text* components tokens are defined with raw and semantic tokens of sizes (from `AllSizeSemanticTokensProvider`), colors (`AllColorSemanticTokensProvider`)
 /// and spaces (from `AllSpaceSemanticTokensProvider`).
 ///
 /// - Since: 0.17.0
-open class OrangeInverseThemeInputTextComponentTokensProvider: AllInputTextComponentTokensProvider {
+final class OrangeInverseThemeTextInputComponentTokensProvider: AllTextInputComponentTokensProvider {
 
-    /// Provider of size semantic tokens to use for button sizes
+    /// Provider of size semantic tokens to use for text input sizes
     public let sizes: AllSizeSemanticTokensProvider
 
-    /// Provider of color semantic tokens to use for button colors
+    /// Provider of border semantic tokens to use for text input sborders
+    public let borders: AllBorderSemanticTokensProvider
+
+    /// Provider of color semantic tokens to use for  text input colors
     public let colors: AllColorSemanticTokensProvider
 
-    /// Provider of spaces semantic tokens to use for button spaces
+    /// Provider of spaces semantic tokens to use for  text input spaces
     public let spaces: AllSpaceSemanticTokensProvider
 
     #if DEBUG
@@ -43,19 +46,22 @@ open class OrangeInverseThemeInputTextComponentTokensProvider: AllInputTextCompo
     /// Defines a provider of component tokens dedicated to `OUDSButton`
     /// - Parameters:
     ///    - sizes: Provider for size semantic tokens. If nil, a default one will be used (`OrangeInverseThemeSizeSemanticTokensProvider`)
+    ///    - borders: Provider for borders semantic tokens. If nil, a default one will be used (`OrangeInverseThemeBorderSemanticTokensProvider`)
     ///    - colors: Provider for color semantic tokens. If nil, a default one will be used (`OrangeInverseThemeColorSemanticTokensProvider`)
     ///    - spaces: Provider for space semantic tokens. If nil, a default one will be used (`OrangeInverseThemeSpaceSemanticTokensProvider`)
     public init(sizes: AllSizeSemanticTokensProvider? = nil,
+                borders: AllBorderSemanticTokensProvider? = nil,
                 colors: AllColorSemanticTokensProvider? = nil,
                 spaces: AllSpaceSemanticTokensProvider? = nil)
     {
-        OL.debug("Init of OrangeInverseThemeInputTextComponentTokensProvider")
+        OL.debug("Init of OrangeInverseThemeTextInputComponentTokensProvider")
         self.sizes = (sizes ?? OrangeInverseThemeSizeSemanticTokensProvider())
+        self.borders = (borders ?? OrangeInverseThemeBorderSemanticTokensProvider())
         self.colors = (colors ?? OrangeInverseThemeColorSemanticTokensProvider())
         self.spaces = (spaces ?? OrangeInverseThemeSpaceSemanticTokensProvider())
         #if DEBUG
         Self.instanceCount++
-        checkInstances(count: Self.instanceCount, for: "OrangeInverseThemeInputTextComponentTokensProvider")
+        checkInstances(count: Self.instanceCount, for: "OrangeInverseThemeTextInputComponentTokensProvider")
         #endif
     }
 
@@ -67,8 +73,8 @@ open class OrangeInverseThemeInputTextComponentTokensProvider: AllInputTextCompo
 
     // ଘ( ･ω･)_/ﾟ･:*:･｡☆
     // Note: So as to help the integration of generated code produced by the tokenator
-    // the implemention of InputTextComponentTokens is not here but in Core/Themes/OrangeInverse/Values/ComponentTokens/OrangeInverseTheme+InputTextComponentTokens.swift
-    // This declaration of OrangeInverseThemeInputTextComponentTokensProvider is here also to allow to write documentation.
+    // the implemention of TextInputComponentTokens is not here but in Core/Themes/OrangeInverse/Values/ComponentTokens/OrangeInverseTheme+TextInputComponentTokens.swift
+    // This declaration of OrangeInverseThemeTextInputComponentTokensProvider is here also to allow to write documentation.
 }
 
 // swiftlint:enable type_name