+++ OUDS/Core/Themes/Sosh/Sources/Providers/ComponentTokens/SoshThemeInputTextComponentTokensProvider.swift
@@ -16,46 +16,52 @@ import OUDSFoundations
 
 // swiftlint:disable type_name
 
-/// A class which wraps all **component  tokens of input text** for *input text* objects like `OUDSInputText`.
+/// A class which wraps all **component  tokens of input text** for *input text* objects like `OUDSTextInput`.
 /// Contains also references to semantic tokens providers so as to be able to use them to define the component tokens.
-/// This provider should be integrated as a `AllInputTextComponentTokensProvider` implementation inside `OUDSTheme` so as to provide
+/// This provider should be integrated as a `AllTextInputComponentTokensProvider` implementation inside `OUDSTheme` so as to provide
 /// all tokens to the users. It helps users to override some of the tokens and assign them to an `OUDSTheme` implementation to use.
 /// It implements also the protocol `InouttextComponentTokens` so as to expose the component tokens for *input text* through any `OUDSTheme`.
 /// *Inout text* components tokens are defined with raw and semantic tokens of sizes (from `AllSizeSemanticTokensProvider`), colors (`AllColorSemanticTokensProvider`)
 /// and spaces (from `AllSpaceSemanticTokensProvider`).
 ///
 /// - Since: 0.17.0
-open class SoshThemeInputTextComponentTokensProvider: AllInputTextComponentTokensProvider {
+final class SoshThemeTextInputComponentTokensProvider: AllTextInputComponentTokensProvider {
 
-    /// Provider of size semantic tokens to use for button sizes
+    /// Provider of size semantic tokens to use for  text input sizes
     public let sizes: AllSizeSemanticTokensProvider
 
-    /// Provider of color semantic tokens to use for button colors
+    /// Provider of border semantic tokens to use for text input borders
+    public let borders: AllBorderSemanticTokensProvider
+
+    /// Provider of color semantic tokens to use for  text input colors
     public let colors: AllColorSemanticTokensProvider
 
-    /// Provider of spaces semantic tokens to use for button spaces
+    /// Provider of spaces semantic tokens to use for  text input spaces
     public let spaces: AllSpaceSemanticTokensProvider
 
     #if DEBUG
     private nonisolated(unsafe) static var instanceCount: Int = 0
     #endif
 
-    /// Defines a provider of component tokens dedicated to `OUDSButton`
+    /// Defines a provider of component tokens dedicated to `OUDSTextInput
     /// - Parameters:
     ///    - sizes: Provider for size semantic tokens. If nil, a default one will be used (`SoshThemeSizeSemanticTokensProvider`)
+    ///    - borders: Provider for borders semantic tokens. If nil, a default one will be used (`SoshThemeBorderSemanticTokensProvider`)
     ///    - colors: Provider for color semantic tokens. If nil, a default one will be used (`SoshThemeColorSemanticTokensProvider`)
     ///    - spaces: Provider for space semantic tokens. If nil, a default one will be used (`SoshThemeSpaceSemanticTokensProvider`)
     public init(sizes: AllSizeSemanticTokensProvider? = nil,
+                borders: AllBorderSemanticTokensProvider? = nil,
                 colors: AllColorSemanticTokensProvider? = nil,
                 spaces: AllSpaceSemanticTokensProvider? = nil)
     {
-        OL.debug("Init of SoshThemeInputTextComponentTokensProvider")
+        OL.debug("Init of SoshThemeTextInputComponentTokensProvider")
         self.sizes = (sizes ?? SoshThemeSizeSemanticTokensProvider())
+        self.borders = (borders ?? SoshThemeBorderSemanticTokensProvider())
         self.colors = (colors ?? SoshThemeColorSemanticTokensProvider())
         self.spaces = (spaces ?? SoshThemeSpaceSemanticTokensProvider())
         #if DEBUG
         Self.instanceCount++
-        checkInstances(count: Self.instanceCount, for: "SoshThemeInputTextComponentTokensProvider")
+        checkInstances(count: Self.instanceCount, for: "SoshThemeTextInputComponentTokensProvider")
         #endif
     }
 
@@ -67,8 +73,8 @@ open class SoshThemeInputTextComponentTokensProvider: AllInputTextComponentToken
 
     // ଘ( ･ω･)_/ﾟ･:*:･｡☆
     // Note: So as to help the integration of generated code produced by the tokenator
-    // the implemention of InputTextComponentTokens is not here but in Core/Themes/Sosh/Values/ComponentTokens/SoshTheme+InputTextComponentTokens.swift
-    // This declaration of SoshThemeInputTextComponentTokensProvider is here also to allow to write documentation.
+    // the implemention of TextInputComponentTokens is not here but in Core/Themes/Sosh/Values/ComponentTokens/SoshTheme+TextInputComponentTokens.swift
+    // This declaration of SoshThemeTextInputComponentTokensProvider is here also to allow to write documentation.
 }
 
 // swiftlint:enable type_name