+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Text input component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Text area component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Select input component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Quantity input component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
@@ -38,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+- Input text component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Select component tokens (tokens library v1.3.0) (Orange-OpenSource/ouds-ios#875)
 - Orange theme font family raw tokens